numaresourcescheduler: move objectstate utils
move scheduler-specific objectstate utilities in their
scheduler subtree
